requirements.txt
streamlit==1.32.2
pytest>=7.0
numpy<2.0
openai>=1.0.0
pypdf>=4.0.0
faiss-cpu>=1.7.0